Update TableOfContents file structure
