+++ context/app/static/js/components/Detail/TableOfContents/index.js
@@ -1,0 +1,3 @@
+import TableOfContents from './TableOfContents';
+
+export default TableOfContents;
